chore: update wording
