+++ src/components/pages/feedbacks/peer-review/MemberPeerReviewsAction.tsx
@@ -2,7 +2,7 @@ import { DeleteOutlined, EyeOutlined } from '@ant-design/icons'
 import { useDisclosure } from '@dwarvesf/react-hooks'
 import { Col, Modal, notification, Row, Tooltip } from 'antd'
 import { Button } from 'components/common/Button'
-import { MemberPeerReviewStatuses } from 'constants/status'
+import { MemberPeerReviewStatus } from 'constants/status'
 import { useState } from 'react'
 import { PeerPerformanceReviewModal } from '../inbox/peer-review/PeerPerformanceReviewModal'
 import { MemberPeerReviewDetail } from './mockData'
@@ -104,7 +104,7 @@ export const MemberPeerReviewsAction = (props: Props) => {
             icon={<EyeOutlined />}
             onClick={openPreviewDialog}
             disabled={
-              memberPeerReviewDetail.status !== MemberPeerReviewStatuses.DONE
+              memberPeerReviewDetail.status !== MemberPeerReviewStatus.DONE
             }
           />
         </Tooltip>
@@ -117,7 +117,7 @@ export const MemberPeerReviewsAction = (props: Props) => {
             icon={<DeleteOutlined />}
             onClick={confirmDelete}
             disabled={
-              memberPeerReviewDetail.status !== MemberPeerReviewStatuses.DRAFT
+              memberPeerReviewDetail.status !== MemberPeerReviewStatus.DRAFT
             }
           />
         </Tooltip>
