+++ src/components/pages/feedbacks/workload/Actions.tsx
@@ -8,12 +8,16 @@ interface Props {
 
 export const Actions = (props: Props) => {
   const { record } = props
-  console.log(record)
 
   return (
     <Row justify="end" gutter={[8, 8]}>
       <Tooltip title="View">
-        <Button type="text-primary" size="small" icon={<EyeOutlined />} />
+        <Button
+          type="text-primary"
+          size="small"
+          icon={<EyeOutlined />}
+          onClick={() => console.log(record)}
+        />
       </Tooltip>
     </Row>
   )
