+++ src/constants/status.ts
@@ -61,17 +61,31 @@ export const peerReviewServeyStatuses: Record<PeerReviewServeyStatus, string> =
     [PeerReviewServeyStatus.SENT]: 'Sent',
   }
 
-export enum MemberPeerReviewStatuses {
+export enum MemberPeerReviewStatus {
   DONE = 'done',
   SENT = 'sent',
   DRAFT = 'draft',
 }
 
-export const memberPeerReviewStatuses: Record<
-  MemberPeerReviewStatuses,
-  string
-> = {
-  [MemberPeerReviewStatuses.DONE]: 'Done',
-  [MemberPeerReviewStatuses.SENT]: 'Sent',
-  [MemberPeerReviewStatuses.DRAFT]: 'Draft',
+export const memberPeerReviewStatuses: Record<MemberPeerReviewStatus, string> =
+  {
+    [MemberPeerReviewStatus.DONE]: 'Done',
+    [MemberPeerReviewStatus.SENT]: 'Sent',
+    [MemberPeerReviewStatus.DRAFT]: 'Draft',
+  }
+
+export enum WorkloadAverageStatus {
+  ALL_BORING_STUFF = 'all-boring-stuff',
+  NOTHING_NEW = 'nothing-new',
+  NOT_MUCH = 'not-much',
+  FEW_THINGS = 'few-things',
+  A_LOT = 'a-lot',
+}
+
+export const workloadAverageStatuses: Record<WorkloadAverageStatus, string> = {
+  [WorkloadAverageStatus.ALL_BORING_STUFF]: 'All boring stuff',
+  [WorkloadAverageStatus.NOTHING_NEW]: 'Nothing new',
+  [WorkloadAverageStatus.NOT_MUCH]: 'Not much',
+  [WorkloadAverageStatus.FEW_THINGS]: 'Few things',
+  [WorkloadAverageStatus.A_LOT]: 'A lot',
 }
